Migrate remaining FormatToolsTest under formats-api
