+++ components/formats-bsd/test/loci/formats/utests/testng.xml
@@ -108,12 +108,6 @@
           <class name="loci.formats.utests.SixteenBitLosslessJPEG2000Test"/>
         </classes>
     </test>
-    <test name="FormatToolsTest">
-      <groups/>
-      <classes>
-        <class name="loci.formats.utests.FormatToolsTest"/>
-      </classes>
-    </test>
     <test name="MemoizerTest">
       <groups/>
       <classes>
